// src/api/models/AllTimeLeaderboardEntry.ts
/* tslint:disable */
/* eslint-disable */
/**
 * tgs_v2
 * Timeguessr Social API V2
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface AllTimeLeaderboardEntry
 */
export interface AllTimeLeaderboardEntry {
  /**
   *
   * @type {string}
   * @memberof AllTimeLeaderboardEntry
   */
  dailyDate: string;
  /**
   *
   * @type {number}
   * @memberof AllTimeLeaderboardEntry
   */
  score: number;
  /**
   *
   * @type {string}
   * @memberof AllTimeLeaderboardEntry
   */
  initials: string;
  /**
   *
   * @type {string}
   * @memberof AllTimeLeaderboardEntry
   */
  uuid: string;
}

/**
 * Check if a given object implements the AllTimeLeaderboardEntry interface.
 */
export function instanceOfAllTimeLeaderboardEntry(
  value: object,
): value is AllTimeLeaderboardEntry {
  if (!("dailyDate" in value) || value["dailyDate"] === undefined) return false;
  if (!("score" in value) || value["score"] === undefined) return false;
  if (!("initials" in value) || value["initials"] === undefined) return false;
  if (!("uuid" in value) || value["uuid"] === undefined) return false;
  return true;
}

export function AllTimeLeaderboardEntryFromJSON(
  json: any,
): AllTimeLeaderboardEntry {
  return AllTimeLeaderboardEntryFromJSONTyped(json, false);
}

export function AllTimeLeaderboardEntryFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): AllTimeLeaderboardEntry {
  if (json == null) {
    return json;
  }
  return {
    dailyDate: json["dailyDate"],
    score: json["score"],
    initials: json["initials"],
    uuid: json["uuid"],
  };
}

export function AllTimeLeaderboardEntryToJSON(
  json: any,
): AllTimeLeaderboardEntry {
  return AllTimeLeaderboardEntryToJSONTyped(json, false);
}

export function AllTimeLeaderboardEntryToJSONTyped(
  value?: AllTimeLeaderboardEntry | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    dailyDate: value["dailyDate"],
    score: value["score"],
    initials: value["initials"],
    uuid: value["uuid"],
  };
}